Username Validation and Form Changes
Added validation to username field consistent with the password and
email fields.
Changed the register scripts to show/hide specific relevant/irrelevant
fields as necessary (e.g. hide matric no when not choosing dundee uni).
Added terms of use checkbox to the form and a terms of use page &
controller.